Refactor cgroups into separate nmdules
This makes the implementations a little easier to read and other changes
to make it much easier to test certain things.
